Write control characters in compression
Cool, we're already able now to write away (i.e. encode) the control characters in the compression algorithm. We can do this based on a given "best length" and "best offset". So basically, what we need to do now is implementing the sliding window to find this best offset and best length.