Use terrain map for pipes
We now lay out the pipes according to the terrain, but the problem is that the squares are not perfect planes, meaning that game won't allow us to draw over it. To fix this, we need to make sure that for straight sections two coordinates are the same, and for all the rest we need a flat tile. This isn't going to be easy though!