Use dedicated TGI index in DBPF's
Instead of storing all entries of the dbpf in a bare array, we now store it in a fancy array that *acts* like a normaly array, but actually adds some indexing mechanism to index by tgi, even though this is not mandatory. This means that we can choose whenever we want to index. Sweet.