Refactor file index to build in separate threads
We've refactored the file index so that it now parses all dbpfs in a separate thread. On a 16-core system, this yields a significant improvement of the speed in which we parse all the dbpfs, allthough filesystem access is still a bottleneck. At least parsing the dbpf now happens in parallel, which is where the 33% speed increase comes from.