Fix circular dependency references
We've noticed that QueryExemplarGUIDs can have the same IID as the building exemplar. When this happens, we got stuck in a deadlock because we didn't use TypeID 0x00000000 when looking for query exemplars, and hence the building exemplar itself was found, resulting in a circular reference and hence a deadlock. Should be fixed now.