Always check for equality
There was an issue with our binary index because if there's a hash collision between a TGI that is not in the hashmap and a collision that *is* in the hashmap, then the item was returned if there were no other collisions. This is wrong obviously, so we simply always have to check for equality instead, which is what a real hash map does as well.