Automatically adjust size
The size of the hash map is no longer a bottleneck because we no longer use objects for the linked list. The linked list is now built up using Uint32Arrays, which are cheap to construct. Hence we now use the nearest power of two for the size of the hashmap, which drastically speeds up the queries as the collisions are reduced.