Use underlying buffer for dbpf header
Given that we're optimizing for speed as much as possible, we've reimplented the dbpf header class on top of a data view. That way, we only have to parse the data we actually need for parsing a dbpf - which is the index offset. The rest we can leave alone and only need to return just in time.